Fix various data type problems
 - Make sure the mockup constant for AT_FDCWD is actualy an int
 - Don't use %lld printf format specified, mscrt doesn't have that
 - On 64 bit windows, use %I64u format specified for size_t
 - Seperate the overflow macro stuff from the form specifier stuff
 - Move the whole thing to compat.h and clean it up a little so
   it becomes readable.

+++ difftool/super.c
@@ -36,8 +36,8 @@ static void print_value_difference(const char *name, sqfs_u64 a, sqfs_u64 b)
 			c = '-';
 			diff = a - b;
 		}
-		fprintf(stdout, "%s: %c%llu\n", name, c,
-			(unsigned long long)diff);
+		fprintf(stdout, "%s: %c%lu\n", name, c,
+			(unsigned long)diff);
 	}
 }
